chatbot/backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# LangGraph & LangChain
langgraph>=1.0.1
langchain>=0.3.27
langchain-core>=0.3.0
langchain-google-genai>=2.1.12
langchain-community>=0.3.30

# Vector Database & Embeddings
chromadb>=1.2.1
sentence-transformers>=2.3.0

# Web Scraping
beautifulsoup4>=4.12.0
playwright>=1.41.0
requests>=2.31.0

# Database
motor>=3.3.2
pymongo>=4.6.1

# Utilities
python-multipart>=0.0.6